[DCS] Change Capacity to float32 (#380)
[DCS] Change Capacity to float32

What this PR does / why we need it
Change Capacity from int to float32
Which issue this PR fixes
fixes #1839

+++ openstack/dcs/v1/instances/results.go
@@ -36,7 +36,7 @@ type ListDcsResponse struct {
 type Instance struct {
 	Name                 string               `json:"name"`
 	Engine               string               `json:"engine"`
-	Capacity             int                  `json:"capacity"`
+	Capacity             float32              `json:"capacity"`
 	IP                   string               `json:"ip"`
 	Port                 int                  `json:"port"`
 	Status               string               `json:"status"`
